Move uTP packet tests to packets.rs file
