+++ trin-core/src/utp/packets.rs
@@ -498,3 +498,429 @@ impl fmt::Debug for Packet {
             .finish()
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::utp::packets::PacketType::{Data, State};
+    use crate::utp::packets::*;
+    use crate::utp::time::{Delay, Timestamp};
+    use quickcheck::{QuickCheck, TestResult};
+    use std::convert::TryFrom;
+
+    #[test]
+    fn test_decode_packet() {
+        let buf = [
+            0x21, 0x0, 0xA4, 0x46, 0xA7, 0x3E, 0xF4, 0x40, 0x0, 0x0, 0x27, 0x10, 0x0, 0x0, 0xF0,
+            0x0, 0x3C, 0x2C, 0x7E, 0xB5,
+        ];
+
+        let packet = Packet::try_from(&buf[..]);
+        assert!(packet.is_ok());
+        let packet = packet.unwrap();
+        assert_eq!(packet.get_type(), PacketType::State);
+        assert_eq!(packet.get_version(), VERSION);
+        assert_eq!(packet.get_extension_type(), ExtensionType::None);
+        assert_eq!(packet.connection_id(), 42054);
+        assert_eq!(packet.timestamp(), 2805920832.into());
+        assert_eq!(packet.timestamp_difference(), 10000.into());
+        assert_eq!(packet.wnd_size(), 61440);
+        assert_eq!(packet.seq_nr(), 15404);
+        assert_eq!(packet.ack_nr(), 32437);
+        assert_eq!(packet.len(), buf.len());
+        assert!(packet.payload().is_empty());
+    }
+
+    #[test]
+    fn test_decode_packet_with_extension() {
+        let buf = [
+            0x21, 0x1, 0xA4, 0x46, 0xA7, 0x3E, 0xF4, 0x40, 0x0, 0x0, 0x27, 0x10, 0x0, 0x0, 0xF0,
+            0x0, 0x3C, 0x2C, 0x7E, 0xB5, 0x0, 0x4, 0x00, 0x00, 0x00, 0x00,
+        ];
+
+        let packet = Packet::try_from(&buf[..]);
+        assert!(packet.is_ok());
+
+        let packet = packet.unwrap();
+        assert_eq!(packet.get_type(), PacketType::State);
+        assert_eq!(packet.get_version(), VERSION);
+        assert_eq!(packet.get_extension_type(), ExtensionType::SelectiveAck);
+        assert_eq!(packet.connection_id(), 42054);
+        assert_eq!(packet.timestamp(), 2805920832.into());
+        assert_eq!(packet.timestamp_difference(), 10000.into());
+        assert_eq!(packet.wnd_size(), 61440);
+        assert_eq!(packet.seq_nr(), 15404);
+        assert_eq!(packet.ack_nr(), 32437);
+        assert_eq!(packet.len(), buf.len());
+        assert!(packet.payload().is_empty());
+        let extensions: Vec<Extension<'_>> = packet.extensions().collect();
+        assert_eq!(extensions.len(), 1);
+        assert_eq!(extensions[0].ty, ExtensionType::SelectiveAck);
+        assert_eq!(extensions[0].data, &[0, 0, 0, 0]);
+        assert_eq!(extensions[0].len(), extensions[0].data.len());
+        assert_eq!(extensions[0].len(), 4);
+        // Reversible
+        assert_eq!(packet.as_ref(), &buf);
+    }
+
+    #[test]
+    fn test_packet_decode_with_missing_extension() {
+        let buf = [
+            0x21, 0x01, 0x41, 0xa8, 0x99, 0x2f, 0xd0, 0x2a, 0x9f, 0x4a, 0x26, 0x21, 0x00, 0x10,
+            0x00, 0x00, 0x3a, 0xf2, 0x6c, 0x79,
+        ];
+        let packet = Packet::try_from(&buf[..]);
+        assert!(packet.is_err());
+    }
+
+    #[test]
+    fn test_packet_decode_with_malformed_extension() {
+        let buf = [
+            0x21, 0x01, 0x41, 0xa8, 0x99, 0x2f, 0xd0, 0x2a, 0x9f, 0x4a, 0x26, 0x21, 0x00, 0x10,
+            0x00, 0x00, 0x3a, 0xf2, 0x6c, 0x79, 0x00, 0x04, 0x00,
+        ];
+        let packet = Packet::try_from(&buf[..]);
+        assert!(packet.is_err());
+    }
+
+    #[test]
+    fn test_decode_packet_with_unknown_extensions() {
+        let buf = [
+            0x21, 0x01, 0x41, 0xa7, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+            0x05, 0xdc, 0xab, 0x53, 0x3a, 0xf5, 0xff, 0x04, 0x00, 0x00, 0x00,
+            0x00, // Imaginary extension
+            0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
+        ];
+        match Packet::try_from(&buf[..]) {
+            Ok(packet) => {
+                assert_eq!(packet.get_version(), 1);
+                assert_eq!(packet.get_extension_type(), ExtensionType::SelectiveAck);
+                assert_eq!(packet.get_type(), State);
+                assert_eq!(packet.connection_id(), 16807);
+                assert_eq!(packet.timestamp(), Timestamp(0));
+                assert_eq!(packet.timestamp_difference(), Delay(0));
+                assert_eq!(packet.wnd_size(), 1500);
+                assert_eq!(packet.seq_nr(), 43859);
+                assert_eq!(packet.ack_nr(), 15093);
+                assert!(packet.payload().is_empty());
+                // The invalid extension is discarded
+                let extensions: Vec<Extension<'_>> = packet.extensions().collect();
+                assert_eq!(extensions.len(), 2);
+                assert_eq!(extensions[0].ty, ExtensionType::SelectiveAck);
+                assert_eq!(extensions[0].data, &[0, 0, 0, 0]);
+                assert_eq!(extensions[0].len(), extensions[0].data.len());
+                assert_eq!(extensions[0].len(), 4);
+            }
+            Err(ref e) => panic!("{}", e),
+        }
+    }
+
+    #[test]
+    fn test_encode_packet() {
+        let payload = b"Hello\n".to_vec();
+        let timestamp = Timestamp(15270793);
+        let timestamp_diff = Delay(1707040186);
+        let (connection_id, seq_nr, ack_nr): (u16, u16, u16) = (16808, 15090, 17096);
+        let window_size: u32 = 1048576;
+        let mut packet = Packet::with_payload(&payload[..]);
+        packet.set_type(Data);
+        packet.set_timestamp(timestamp);
+        packet.set_timestamp_difference(timestamp_diff);
+        packet.set_connection_id(connection_id);
+        packet.set_seq_nr(seq_nr);
+        packet.set_ack_nr(ack_nr);
+        packet.set_wnd_size(window_size);
+        let buf = [
+            0x01, 0x00, 0x41, 0xa8, 0x00, 0xe9, 0x03, 0x89, 0x65, 0xbf, 0x5d, 0xba, 0x00, 0x10,
+            0x00, 0x00, 0x3a, 0xf2, 0x42, 0xc8, 0x48, 0x65, 0x6c, 0x6c, 0x6f, 0x0a,
+        ];
+
+        assert_eq!(packet.len(), buf.len());
+        assert_eq!(packet.len(), HEADER_SIZE + payload.len());
+        assert_eq!(&packet.payload(), &payload.as_slice());
+        assert_eq!(packet.get_version(), 1);
+        assert_eq!(packet.get_extension_type(), ExtensionType::None);
+        assert_eq!(packet.get_type(), Data);
+        assert_eq!(packet.connection_id(), connection_id);
+        assert_eq!(packet.seq_nr(), seq_nr);
+        assert_eq!(packet.ack_nr(), ack_nr);
+        assert_eq!(packet.wnd_size(), window_size);
+        assert_eq!(packet.timestamp(), timestamp);
+        assert_eq!(packet.timestamp_difference(), timestamp_diff);
+        assert_eq!(packet.as_ref(), buf);
+    }
+
+    #[test]
+    fn test_encode_packet_with_payload() {
+        let payload = b"Hello\n".to_vec();
+        let timestamp = Timestamp(15270793);
+        let timestamp_diff = Delay(1707040186);
+        let (connection_id, seq_nr, ack_nr): (u16, u16, u16) = (16808, 15090, 17096);
+        let window_size: u32 = 1048576;
+        let mut packet = Packet::with_payload(&payload[..]);
+        packet.set_timestamp(timestamp);
+        packet.set_timestamp_difference(timestamp_diff);
+        packet.set_connection_id(connection_id);
+        packet.set_seq_nr(seq_nr);
+        packet.set_ack_nr(ack_nr);
+        packet.set_wnd_size(window_size);
+        let buf = [
+            0x01, 0x00, 0x41, 0xa8, 0x00, 0xe9, 0x03, 0x89, 0x65, 0xbf, 0x5d, 0xba, 0x00, 0x10,
+            0x00, 0x00, 0x3a, 0xf2, 0x42, 0xc8, 0x48, 0x65, 0x6c, 0x6c, 0x6f, 0x0a,
+        ];
+
+        assert_eq!(packet.len(), buf.len());
+        assert_eq!(packet.len(), HEADER_SIZE + payload.len());
+        assert_eq!(&packet.payload(), &payload.as_slice());
+        assert_eq!(packet.get_version(), 1);
+        assert_eq!(packet.get_type(), Data);
+        assert_eq!(packet.get_extension_type(), ExtensionType::None);
+        assert_eq!(packet.connection_id(), connection_id);
+        assert_eq!(packet.seq_nr(), seq_nr);
+        assert_eq!(packet.ack_nr(), ack_nr);
+        assert_eq!(packet.wnd_size(), window_size);
+        assert_eq!(packet.timestamp(), timestamp);
+        assert_eq!(packet.timestamp_difference(), timestamp_diff);
+        assert_eq!(packet.as_ref(), buf);
+    }
+
+    #[test]
+    fn test_reversible() {
+        let buf = [
+            0x01, 0x00, 0x41, 0xa8, 0x00, 0xe9, 0x03, 0x89, 0x65, 0xbf, 0x5d, 0xba, 0x00, 0x10,
+            0x00, 0x00, 0x3a, 0xf2, 0x42, 0xc8, 0x48, 0x65, 0x6c, 0x6c, 0x6f, 0x0a,
+        ];
+        assert_eq!(&Packet::try_from(&buf[..]).unwrap().as_ref(), &buf);
+    }
+
+    #[test]
+    fn test_decode_evil_sequence() {
+        let buf = [
+            0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        ];
+        let packet = Packet::try_from(&buf[..]);
+        assert!(packet.is_err());
+    }
+
+    #[test]
+    fn test_decode_empty_packet() {
+        let packet = Packet::try_from(&[][..]);
+        assert!(packet.is_err());
+    }
+
+    #[test]
+    fn test_packet_set_type() {
+        let mut packet = Packet::new();
+        packet.set_type(PacketType::Syn);
+        assert_eq!(packet.get_type(), PacketType::Syn);
+        packet.set_type(PacketType::State);
+        assert_eq!(packet.get_type(), PacketType::State);
+        packet.set_type(PacketType::Fin);
+        assert_eq!(packet.get_type(), PacketType::Fin);
+        packet.set_type(PacketType::Reset);
+        assert_eq!(packet.get_type(), PacketType::Reset);
+        packet.set_type(PacketType::Data);
+        assert_eq!(packet.get_type(), PacketType::Data);
+    }
+
+    #[test]
+    fn test_packet_set_selective_acknowledgment() {
+        let mut packet = Packet::new();
+        packet.set_selective_ack(vec![1, 2, 3, 4]);
+
+        {
+            let extensions: Vec<Extension<'_>> = packet.extensions().collect();
+            assert_eq!(extensions.len(), 1);
+            assert_eq!(extensions[0].ty, ExtensionType::SelectiveAck);
+            assert_eq!(extensions[0].data, &[1, 2, 3, 4]);
+            assert_eq!(extensions[0].len(), extensions[0].data.len());
+            assert_eq!(extensions[0].len(), 4);
+        }
+
+        // Add a second sack
+        packet.set_selective_ack(vec![5, 6, 7, 8, 9, 10, 11, 12]);
+
+        let extensions: Vec<Extension<'_>> = packet.extensions().collect();
+        assert_eq!(extensions.len(), 2);
+        assert_eq!(extensions[0].ty, ExtensionType::SelectiveAck);
+        assert_eq!(extensions[0].data, &[1, 2, 3, 4]);
+        assert_eq!(extensions[0].len(), extensions[0].data.len());
+        assert_eq!(extensions[0].len(), 4);
+        assert_eq!(extensions[1].ty, ExtensionType::SelectiveAck);
+        assert_eq!(extensions[1].data, &[5, 6, 7, 8, 9, 10, 11, 12]);
+        assert_eq!(extensions[1].len(), extensions[1].data.len());
+        assert_eq!(extensions[1].len(), 8);
+    }
+
+    // Use quickcheck to simulate a malicious attacker sending malformed packets
+    #[test]
+    fn quicktest() {
+        fn run(x: Vec<u8>) -> TestResult {
+            let packet = Packet::try_from(x.as_slice());
+
+            if PacketHeader::try_from(x.as_slice())
+                .and(check_extensions(x.as_slice()))
+                .is_err()
+            {
+                TestResult::from_bool(packet.is_err())
+            } else if let Ok(packet) = packet {
+                TestResult::from_bool(packet.as_ref() == x.as_slice())
+            } else {
+                TestResult::from_bool(false)
+            }
+        }
+        QuickCheck::new()
+            .tests(10000)
+            .quickcheck(run as fn(Vec<u8>) -> TestResult)
+    }
+
+    #[test]
+    fn extension_iterator() {
+        let buf = [
+            0x21, 0x00, 0x41, 0xa8, 0x99, 0x2f, 0xd0, 0x2a, 0x9f, 0x4a, 0x26, 0x21, 0x00, 0x10,
+            0x00, 0x00, 0x3a, 0xf2, 0x6c, 0x79,
+        ];
+        let packet = Packet::try_from(&buf[..]).unwrap();
+        assert_eq!(packet.extensions().count(), 0);
+
+        let buf = [
+            0x21, 0x01, 0x41, 0xa7, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+            0x05, 0xdc, 0xab, 0x53, 0x3a, 0xf5, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
+        ];
+        let packet = Packet::try_from(&buf[..]).unwrap();
+        let extensions: Vec<Extension<'_>> = packet.extensions().collect();
+        assert_eq!(extensions.len(), 1);
+        assert_eq!(extensions[0].ty, ExtensionType::SelectiveAck);
+        assert_eq!(extensions[0].data, &[0, 0, 0, 0]);
+        assert_eq!(extensions[0].len(), extensions[0].data.len());
+        assert_eq!(extensions[0].len(), 4);
+
+        let buf = [
+            0x21, 0x01, 0x41, 0xa7, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+            0x05, 0xdc, 0xab, 0x53, 0x3a, 0xf5, 0xff, 0x04, 0x01, 0x02, 0x03,
+            0x04, // Imaginary extension
+            0x00, 0x04, 0x05, 0x06, 0x07, 0x08,
+        ];
+
+        let packet = Packet::try_from(&buf[..]).unwrap();
+        let extensions: Vec<Extension<'_>> = packet.extensions().collect();
+        assert_eq!(extensions.len(), 2);
+        assert_eq!(extensions[0].ty, ExtensionType::SelectiveAck);
+        assert_eq!(extensions[0].data, &[1, 2, 3, 4]);
+        assert_eq!(extensions[0].len(), extensions[0].data.len());
+        assert_eq!(extensions[0].len(), 4);
+        assert_eq!(extensions[1].ty, ExtensionType::Unknown(0xff));
+        assert_eq!(extensions[1].data, &[5, 6, 7, 8]);
+        assert_eq!(extensions[1].len(), extensions[1].data.len());
+        assert_eq!(extensions[1].len(), 4);
+    }
+
+    // https://github.com/ethereum/portal-network-specs/pull/127
+    // konrad uTP test vectors
+
+    #[test]
+    fn test_syn_packet() {
+        let mut packet = Packet::new();
+        packet.set_type(PacketType::Syn);
+        packet.set_connection_id(10049);
+        packet.set_timestamp(3384187322.into());
+        packet.set_timestamp_difference(0.into());
+        packet.set_wnd_size(1048576);
+        packet.set_seq_nr(11884);
+        packet.set_ack_nr(0);
+
+        println!("packet: {:?}", packet);
+        println!("packet raw: {:?}", packet.as_ref());
+
+        assert_eq!(
+            hex::encode(packet.as_ref()),
+            "41002741c9b699ba00000000001000002e6c0000"
+        );
+    }
+
+    #[test]
+    fn test_act_packet_no_extension() {
+        let mut packet = Packet::new();
+        packet.set_type(PacketType::State);
+        packet.set_connection_id(10049);
+        packet.set_timestamp(6195294.into());
+        packet.set_timestamp_difference(916973699.into());
+        packet.set_wnd_size(1048576);
+        packet.set_seq_nr(16807);
+        packet.set_ack_nr(11885);
+
+        assert_eq!(
+            hex::encode(packet.as_ref()),
+            "21002741005e885e36a7e8830010000041a72e6d"
+        );
+    }
+
+    #[test]
+    fn test_act_packet_with_selective_ack_extension() {
+        let mut packet = Packet::new();
+        packet.set_type(PacketType::State);
+        packet.set_connection_id(10049);
+        packet.set_timestamp(6195294.into());
+        packet.set_timestamp_difference(916973699.into());
+        packet.set_wnd_size(1048576);
+        packet.set_seq_nr(16807);
+        packet.set_ack_nr(11885);
+
+        packet.set_selective_ack(vec![1, 0, 0, 128]);
+
+        assert_eq!(
+            hex::encode(packet.as_ref()),
+            "21012741005e885e36a7e8830010000041a72e6d000401000080"
+        );
+    }
+
+    #[test]
+    fn test_data_packet() {
+        let mut packet = Packet::with_payload(&[0, 1, 2, 3, 4, 5, 6, 7, 8, 9]);
+        packet.set_type(PacketType::Data);
+        packet.set_connection_id(26237);
+        packet.set_timestamp(252492495.into());
+        packet.set_timestamp_difference(242289855.into());
+        packet.set_wnd_size(1048576);
+        packet.set_seq_nr(8334);
+        packet.set_ack_nr(16806);
+
+        assert_eq!(
+            hex::encode(packet.as_ref()),
+            "0100667d0f0cbacf0e710cbf00100000208e41a600010203040506070809"
+        );
+    }
+
+    #[test]
+    fn test_fin_packet() {
+        let mut packet = Packet::new();
+        packet.set_type(PacketType::Fin);
+        packet.set_connection_id(19003);
+        packet.set_timestamp(515227279.into());
+        packet.set_timestamp_difference(511481041.into());
+        packet.set_wnd_size(1048576);
+        packet.set_seq_nr(41050);
+        packet.set_ack_nr(16806);
+
+        assert_eq!(
+            hex::encode(packet.as_ref()),
+            "11004a3b1eb5be8f1e7c94d100100000a05a41a6"
+        );
+    }
+
+    #[test]
+    fn test_reset_packet() {
+        let mut packet = Packet::new();
+        packet.set_type(PacketType::Reset);
+        packet.set_connection_id(62285);
+        packet.set_timestamp(751226811.into());
+        packet.set_timestamp_difference(0.into());
+        packet.set_wnd_size(0);
+        packet.set_seq_nr(55413);
+        packet.set_ack_nr(16807);
+
+        assert_eq!(
+            hex::encode(packet.as_ref()),
+            "3100f34d2cc6cfbb0000000000000000d87541a7"
+        );
+    }
+}
